:recycle: refactor: http cache controller and object seperated
